backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
reportlab==4.2.0
requests==2.32.2
aiofiles==23.2.1
pydantic[email]==2.7.1
gitpython==3.1.43
bandit==1.8.3
